+++ CHANGELOG.md
@@ -6,7 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.5.8] - 2024-01-30
+
+### Added
+
+- Add Ingestion sources
+
+## [0.5.7] - 2024-01-30
+
+### Added
+
+- Update tsconfig, babel config and package name
 
 ## [0.5.6] - 2023-08-30
 
@@ -44,7 +54,7 @@ and this project adheres to
 
 ### Fixed
 
-- Possible duplicate relationships between site and scan entities are now 
+- Possible duplicate relationships between site and scan entities are now
   properly handled.
 
 ## [0.5.0] - 2023-08-01